Optimize Osu class for improved performance
The get_key_subset function has been modified to return a reference to a statically-allocated string instead of allocating a new string on the heap. This avoids an unnecessary memory allocation and copy.

The get_key_subset function has been modified to use a loop to copy each character individually instead of using multiple memory copies with memmove_s. This may improve the performance of the function in some cases.

The get_key_subset function now uses an assertion to check the value of column_count instead of throwing an exception. This avoids the overhead of throwing and catching an exception.

The Osu constructor has been modified to run the memory reads in parallel using std::async. This may improve the performance of the constructor by allowing the reads to be performed concurrently.

The get_map_player method has been modified to cache the map_player object to avoid unnecessary memory reads.

The log method has been left in the release build and is guarded by an #ifdef DEBUG preprocessor directive to enable the compiler to optimize it out in release builds. This may improve the performance of the method in release builds.